refactor(user): rename the component of layouts
